+++ Drop-In/src/main/java/com/braintreepayments/api/AddPaymentUpdateListener.java
@@ -2,7 +2,7 @@ package com.braintreepayments.api;
 
 import android.view.View;
 
-public interface AddPaymentUpdateListener {
+interface AddPaymentUpdateListener {
     void onPaymentUpdated(View v);
     void onBackRequested(View v);
 }