+++ Drop-In/src/main/java/com/braintreepayments/api/BaseActivity.java
@@ -15,7 +15,7 @@ import org.json.JSONException;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-public class BaseActivity extends AppCompatActivity {
+class BaseActivity extends AppCompatActivity {
 
     static final String EXTRA_CONFIGURATION_DATA = "com.braintreepayments.api.EXTRA_CONFIGURATION_DATA";