+++ Drop-In/src/main/java/com/braintreepayments/api/EditCardView.java
@@ -10,6 +10,8 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.LinearLayout;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.braintreepayments.api.dropin.R;
 import com.braintreepayments.api.exceptions.BraintreeError;
 import com.braintreepayments.api.exceptions.ErrorWithResponse;
@@ -20,9 +22,7 @@ import com.braintreepayments.cardform.OnCardFormSubmitListener;
 import com.braintreepayments.cardform.view.CardEditText;
 import com.braintreepayments.cardform.view.CardForm;
 
-import androidx.appcompat.app.AppCompatActivity;
-
-public class EditCardView extends LinearLayout implements OnCardFormFieldFocusedListener, OnClickListener,
+class EditCardView extends LinearLayout implements OnCardFormFieldFocusedListener, OnClickListener,
         OnCardFormSubmitListener {
 
     private CardForm mCardForm;