+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android Drop-In Release Notes
 
+## unreleased
+
+* Breaking Changes
+  * Make `AddCardView` package-private
+
 ## 5.2.0
 
 * Update android-card-form version to 5.1.0 (fixes #212)