+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Breaking Changes
   * Make `AddCardView` package-private
   * Make `AddPaymentUpdateListener` package-private
+  * Make `AnimatedButtonView` package-private
 
 ## 5.2.0