+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Make `EditCardView` package-private
   * Make `EnrollmentCardView` package-private
   * Make `PaymentMethodItemView` package-private
+  * Make `SupportedPaymentMethodsAdapter` package-private
 
 ## 5.2.0